Rearranged Code. Split bigger classes into smaller ones
