+++ Assets/Scripts/Classes/EntityWithTime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0c446e06fe1b8436c9b70d24e17791be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
