+++ Assets/Scripts/Classes/PlainCheckList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 61c379e7d76be4052a61f5066c3f8749
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
