+++ Assets/Scripts/Classes/PlainCheckListItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2378a695540c143ddbffaa1b322a68cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
