+++ Assets/Scripts/Managers/PlayerManager.cs
@@ -6,6 +6,18 @@ public class PlayerManager : MonoBehaviour
     public static event Action OnPlayerStartedGame;
     public static event Action OnPlayerEndedGame;
 
+    private static GameObject _player;
+
+    private void Start()
+    {
+       FetchPlayer();
+    }
+
+    private static void FetchPlayer()
+    {
+        _player = GameObject.FindGameObjectWithTag("Player");
+    }
+
     private void OnTriggerEnter(Collider other)
     {
         if (other.gameObject.CompareTag("StartGame") && !GameManager.GameStarted())
@@ -24,5 +36,11 @@ public class PlayerManager : MonoBehaviour
                 return;
             }
         }
+    }
+
+    public static GameObject GetPlayer()
+    {
+        if (!_player) FetchPlayer();
+        return _player;
     }
 }
